Backport of the pre-0.77 demo features to 0.76.
This is a combined cherry-pick of three commits on the 0.77 branch:

896bcd506866d3f "Resurrect the test backends"
9294ee349617632 "Windows PuTTYgen: saw load_key_file in half"
bc7e06c49411a89 "Windows tools: assorted '-demo' options"

Of course, these commits aren't in 0.76 itself, and I don't see any
reason to publish a binary build of 0.76 containing them. This is just
a branch I can build from myself to get a version of PuTTY that will
look exactly like the real 0.76 but permit automated generation of
demo screenshots.